+++ internal/handler/sign.go
@@ -194,6 +194,8 @@ func (h *SignHandler) FinishSignData(c *gin.Context) {
 		c.Error(err)
 		return
 	}
-	resp.Data = res
+	resp.Data = gin.H{
+		"payId": res,
+	}
 	c.JSON(http.StatusOK, resp)
 }